mysite/requirements.txt
asgiref==3.4.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
click==7.1.2
colorama==0.4.4
Django==3.2.7
djangorestframework==3.12.4
google-api-core==1.31.3
google-api-python-client==1.12.8
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-cloud-vision==2.4.2
Google-Images-Search==1.3.9
googleapis-common-protos==1.53.0
grpcio==1.40.0
httplib2==0.19.1
idna==3.2
packaging==21.0
Pillow==8.3.2
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyfiglet==0.8.post1
pyparsing==2.4.7
python-resize-image==1.1.19
pytz==2021.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.2
termcolor==1.1.0
uritemplate==3.0.1
urllib3==1.26.6